[버그수정]

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/controller/LikeablePersonController.java
@@ -63,7 +63,7 @@ public class LikeablePersonController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Integer id) {
+    public String delete(@PathVariable("id") Long id) {
         RsData<LikeablePerson> deleteRsData = likeablePersonService.delete(id);
 
         if (deleteRsData.isFail()) {
