+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/service/LikeablePersonService.java
@@ -55,7 +55,7 @@ public class LikeablePersonService {
     }
 
     @Transactional
-    public RsData<LikeablePerson> delete(Integer id) {
+    public RsData<LikeablePerson> delete(Long id) {
         Optional<LikeablePerson> person = likeablePersonRepository.findById(id);
 
         if (person.isEmpty()) {
